refactor: Update npm dependencies for fontawesome-svg-core, free-brands-svg-icons, free-regular-svg-icons, free-solid-svg-icons, and react-fontawesome
This commit updates the npm dependencies for the fontawesome-svg-core, free-brands-svg-icons, free-regular-svg-icons, free-solid-svg-icons, and react-fontawesome packages to version 6.5.2. This update ensures that the latest stable versions of these packages are used in the codebase, which may include bug fixes, performance improvements, or new features. Keeping dependencies up to date is important for maintaining a secure and efficient codebase.