refactor: Add roles to user object in login response
This commit modifies the login function in the auth.controller.js file to include the roles property in the response JSON. Previously, only the user's id was included in the response. By adding the roles property, the client can now access the user's roles after a successful login. This change improves the functionality and flexibility of the login endpoint.