Enable refresh token feature
This commit updates the API routes and hooks to improve consistency and readability. It modifies the `api-routes.ts` file to use consistent indentation and adds the `createRefreshTokenUrl` route. It also adds the `refresh-token.hook.tsx` file to handle refreshing the access token. Additionally, it removes the unused `use-api.ts` file and adds the `axios-private.hook.tsx` file for handling private API requests. These changes enhance the maintainability and functionality of the codebase.